fix(admin): Keep role selection when removing user
This just keeps the role attribute around after a user has been deleted if done by filtering on a role.

Note that it will show a blank table if no more users exist in that group, we could add a check if we should then go to "all" or not. Happy to hear suggestions