Build/Test Tools: Add input for Gutenberg version to E2E workflow.
The E2E test suite has some basic tests to confirm that the plugin can be activated without conflicts. This works great until the minimum version of WordPress required to run the plugin is raised. Old branches that attempt to run these tests that are newly incompatible with the latest version of the plugin will fail.

This change adds a `gutenberg-version` input to the E2E GitHub Action workflow to allow these tests to continue running with the last compatible version.

Props swissspidy.
See #61530.

git-svn-id: https://develop.svn.wordpress.org/trunk@59220 602fd350-edb4-49c9-b593-d223f7449a82