Proceed past BODY and HTML close tags
The BODY and HTML close tag handling in the HTML processor is somewhat
unique in that the token is processed but it only changes internal
state. No elements are pushed or popped, the stack of open elements is
unchanged, and `next_token` must call itself recursively in order to add
something to the element queue.

Remove the `return true` from the close tag handling so that the
processor proceeds until it has something to stop on in `next_token` in
these cases.