I18N: localize `jS` to day+ordinal in `date_i18n`
The built-in PHP `NumberFormatter` can localize numbers to include
ordinal suffixes; however the return value always includes both the
numeric value and also the suffix -- e.g. it is not currently
possible to request only the suffix.

In the case of date formatting, it's fairly typical that if
ordinals are expected, then they will be output for the calendar
day number (e.g. the `11` in `2024-01-11`) - and that the suffix
should appear directly adjacent (e.g. `11th`).

Day-of-month should be restricted to values no larger than `31`, and
should all be integer values, so we shouldn't need to worry about any
locale-specific introduction of thousands or decimal place markers.

Based on the above, we can implement limited/restricted localization
of day-number-plus-ordinal-suffix in the WordPress `date_i18n`
function by special-case detecting an `S` (ordinal format) character
symbol immediately preceded by a `j` (day of the month) format code.

Refs:
- https://www.php.net/manual/en/class.numberformatter.php
- https://www.php.net/manual/en/datetime.format.php

Closes: https://core.trac.wordpress.org/ticket/22225