Passing int term term_exists parent param not respected
This PR fixes a bug in the term_exists function where the parent_term wasn’t being considered correctly. When a term ID (as an integer) was passed along with a parent term ID, the function would only check if the term ID existed and completely ignore whether it was under the specified parent.

Updated the term_exists logic to include the parent in the query when both a term ID and parent term ID are provided.
Added a condition to ensure that the parent is checked as part of the query when relevant.