sanitize_file_name(): Normalize all space characters to a space. (#9103)
Resolves Core-62995

The `sanitize_file_name()` function normalizes the no-break space
to a normal space (U+0020) in order to prevent issues saving files
with the no-break space in it.

This patch expands the replacement to all space characters, since
it’s known that macOS stores a NARROW NO-BREAK SPACE (U+202F) in
screenshot filenames between the time and the am/pm indicator.

There are deeper issues with the way this function works, but this
patch resolves a known and common problem without raising any of
the deeper refactoring questions.