HTML API: Refactor `wp_kses_hair()` (#9248)
Trac ticket: Core-63694

`wp_kses_hair()` is built around an impressive state machine for parsing
the `$attr` of an HTML tag, that is, the span of text after the tag name
and before the closing `>`. Unfortunately, that parsing code doesn’t
fully-implement the HTML specification and may be prone to mis-parsing.

This patch replaces the existing state machine with a straight-forward
use of the HTML API to parse the attributes for us, constructing a shell
take for the `$attr` string and reading the attributes structurally.
This shell is necessary because a previous stage of the pipeline has
already separated what it thinks is the so-called “attribute list” from
a tag.

Props: adamziel, dmsnell, jonsurrell, jorbin, westonruter.

Co-authored-by: Adam Zieliński <zieladam@git.wordpress.org>
Co-authored-by: Aaron Jorbin <jorbin@git.wordpress.org>
Co-authored-by: Jon Surrell <jonsurrell@git.wordpress.org>
CO-authored-by: Weston Ruter <westonruter@git.wordpress.org>
Github-PR: 9248
Github-PR-URL: https://github.com/WordPress/wordpress-develop/pull/9248
Trac-Ticket: 63694
Trac-Ticket-URL: https://core.trac.wordpress.org/ticket/63724
Git-Branch: html-api/refactor-wp-kses-hair-take-3