External Libraries: Adopt standard get_temp_dir() in Text_Diff
This is placing files in a potentially different place than the
rest of WordPress, and its potential false return value was not
checked by the only caller in Text_Diff_Engine_shell.

I considered changing our get_temp_dir implementations to check
the additional directories that Text_Diff::_getTempDir considered,
but decided against it. The Text_Diff code predates the introduction
of sys_get_temp_dir in PHP 5.2, and so is checking these to make up
for that. The latest version of the code that this was based on no
longer does this and calls sys_get_temp_dir instead. Specifically, it
removed _getTempDir() in favor of tempnam(null) [1], and later
switched to horde/util, [2] which uses sys_get_temp_dir. [3]

[1]: https://github.com/horde/Text_Diff/commit/323c4783b9
[2]: https://github.com/horde/Text_Diff/commit/152f6aacf8
[3]: https://github.com/horde/Util/blob/v2.5.12/lib/Horde/Util.php#L200