Add `wp_is_valid_utf8()` for normalizing UTF-8 checks. (#9317)
There are several existing mechanisms in Core to determine if a given
string contains valid UTF-8 bytes or not. These are spread out and
depend on which extensions are installed on the running system. The
`seems_utf8()` method doesn’t validate either; it takes a wild guess
at whether the string looks like UTF-8, ignoring all the rules.

This new method is an efficient and spec-compliant UTF-8 validator.
It depends on no modules except for the optional use of
`mb_check_encoding()`, which holds a special role in PHP 8.3.0+.