Performance: Optimize ASCII spans in `seems_utf8()`. (#9325)
This patch introduces a speedup for the `seems_utf8()` function.
Currently, the function scans every byte in user-land PHP code.
This is prohibitively slow. In this change, spans of ASCII bytes
are fast-tracked through the use of `strspn()`, which can scan
valid UTF-8 single-byte characters at around 3600 MB/s instead of
around 300 MB/s as is the case with the existing code.