Convert two source files to UTF-8
The files in this commit have previously contained text encoded as `Windows-1252` (commonly mistaken for `latin1`/`ISO-8859-1`), in ways that present invalid UTF-8 bytes.

This commit converts the encodings to UTF-8, which resolves the encoding
issues. While the `whole-posts` document could be argued to be
intentionally `Windows-1252`, there was no obvious evidence that this
was more than an historic legacy. The JavaScript file, however,
definitely ought to be UTF-8.