Add a spoken message after delete action, set focus after deletion
When the currently focused control is deleted, a loss of focus is triggered; need to explicitly set the focus. Standard practice is the next item in the collection if there is one. If not, I set it to the previous; if neither, to the tag input.

Also adds wp.a11y.speak after each action.