Simplify a condition in the HTML API's `get_url_in_content` function.
As there's an `is_string()` check already, the `! empty( $href )` can be simplified to a string comparison, as other variable types that are checked in `empty()` won't appear.

`empty()` also returns `false` for the string `"0"` which would however be a valid (relative) URL and thus should be detectable by the function.