Move test method to Imagick specific class
This moves the `test_quality_with_avif_conversion_file_sizes` test method and related filter hooks to the `Tests_Image_Editor_Imagick` class, which specifically forces the use of `WP_Image_Editor_Imagick`.

In the current location (the `Tests_Media` class), a specific image editor is not enforced, so the test can fall back to using GD when Imagick is not compiled with AVIF support.