Charset: Add wp_utf8_chunks() to iterate through strings.
This new generator function iterates through valid and invalid spans
of bytes in a UTF-8 string. It is a convenience wrapper around the
new `_wp_scan_utf8()` which aids a number of operations which should
work on strings containing invalid spans of UTF-8, such as:

 - Identifying where the invalid spans are.
 - Operating on the valid portions while preserving the invalid spans.
 - Displaying broken strings.

These operations can be useful during validation, sanitization,
debugging, and processing uncontrolled inputs which may contain
malformed sequences.