Charset: Track detection of non-characters when scanning UTF-8.
Noncharacters are code points that are permantently reserved in the
Unicode Standard for internal use. They are not recommended for use in
open interchange of Unicode text data. However, they are valid code
points and will not cause a string to return as invalid.

Still, HTML and XML both impose semantic rules on their use and it may
be important for code to know whether they are present in a string. This
patch introduces a new function, `wp_has_noncharacters()`, which answers
this question.

This is accomplished through an inline check with the fallback UTF-8 scanner.
There are 66 noncharacters, making it difficult to find them properly
with common string search functionality. While the inline check adds
overhead to the scanning process, the rare occurrance of noncharacters
should lead to minimal actual overhead due to strong branch prediction.

See https://www.unicode.org/versions/Unicode17.0.0/core-spec/chapter-23/#G12612