Add an exclusion to prevent building full-page of interactivity-router
The full-page wpScriptModuleExport defined in interactivity-router
should not be bundled in Wordpress core yet as it is experimental.
There is no mechanism in wpScriptModuleExports right now that supports
building a script module but not including it in core. Since this
is likely to be a rare instance we have opted to hard-code an
exclusion in core's build until this module is no longer experimental.