Remove invalid `in:body` qualifier
The `in:body` qualifier does not exist in GraphQL, only in GitHub's search syntax.

This adds a `.filter()` before mapping the pull request numbers to the variable that is returned. The filter checks for the presence of the needles within the pull request's body text, which helps filter out any instances where a Trac ticket is referenced within a comment or PR review.