Build/Test Tools: Update MacOS GitHub Actions runner.
The `macos-13` runner image is deprecated with support set to be fully removed next month. This updates the build process testing workflow’s MacOS job to use the `macos-15-intel` runner in the 5.3 branch.

The `-intel` part is required because older branches of WordPress use Node.js `14.x` which never supported `arm64` environments. The `macos-15-intel` image will be the last `x86_64` runner published for GitHub Actions and will be supported until August 2027.

Merges [61226] to the 5.3 branch.

See #64246.

git-svn-id: https://develop.svn.wordpress.org/branches/5.3@61236 602fd350-edb4-49c9-b593-d223f7449a82