Help/About: Update About page header images and performance content.
This adds the correct string for performance changes, updates the header images to not have rounded corners, and updates the release badge to use the stylized 6.9 icon.

Backport includes the fast follow to spell out abbreviation in About page text.

Reviewed by wildworks.
Merges [61257,61258] to the 6.9 branch.

Props ryelle, peterwilsoncc, westonruter, wildworks.
See #63941.


git-svn-id: https://develop.svn.wordpress.org/branches/6.9@61259 602fd350-edb4-49c9-b593-d223f7449a82