Notes: trash (or delete) child notes when parent is deleted.
Ensure that when a top level note is trashed (or deleted), all of its replies (children) are also trashed or deleted. If EMPTY_TRASH_DAYS is 0, notes are deleted immediately; otherwise they are marked as trash for later cleanup.

Reviewed by wildworks.
Merges [61248] to the 6.9 branch.

Props adamsilverstein, desrosj, wildworks, mamaduka, karthickmurugan, jeffpaul, shailu25.
See #64240.

git-svn-id: https://develop.svn.wordpress.org/branches/6.9@61274 602fd350-edb4-49c9-b593-d223f7449a82