Admin: Remove alt syntax from `dashicons.css`
Safari only introduced support for alternative text in generated CSS content in version 17.4, released in 2024. This meets the latest two version declared for core browser support. But because `dashicons.css` is commonly required on the front-end for plugins and themes, this support impacts front-end users, and the browser support guidelines for core should not apply in this context.

Follow up to [60885].

Reviewed by desrosj.
Merges [61290] to the 6.9 branch.

Props codex-m, sabernhardt, wildworks, joedolson, desrosj, welcher.
Fixes #64266.

git-svn-id: https://develop.svn.wordpress.org/branches/6.9@61293 602fd350-edb4-49c9-b593-d223f7449a82